Week_3_Terraform/sales-etl-pipeline/requirements.txt
pandas
apache-airflow
apache-airflow-providers-google
google-cloud-storage
google-cloud-bigquery